Moved mass subystem from external to core
