--- aviary/examples/external_subsystems/dbf_based_mass/__init__.py
@@ -1,1 +1,0 @@
-from .dbf_mass_builder import DBFMassBuilder
